# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lオペ for CLINIC API",
    "description": "クリニック特化LINE運用プラットフォーム — API仕様書",
    "version": "1.0.0",
    "contact": {
      "name": "Lオペ for CLINIC"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/reorder/apply": {
      "post": {
        "summary": "再処方申請",
        "tags": [
          "再処方"
        ],
        "responses": {
          "200": {
            "description": "申請結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "patientCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "patientId": {
                    "type": "string"
                  }
                },
                "required": [
                  "productCode"
                ],
                "additionalProperties": {}
              }
            }
          }
        }
      }
    },
...
    "/api/admin/settings": {
      "get": {
        "summary": "テナント設定取得",
        "tags": [
          "管理: 設定"
        ],
        "responses": {
          "200": {
            "description": "設定値一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      },
      "put": {
        "summary": "テナント設定更新",
        "tags": [
          "管理: 設定"
        ],
        "responses": {
          "200": {
            "description": "更新結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "minLength": 1
                  },
                  "key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "value": {
                    "type": "string"
                  }
                },
                "required": [
                  "category",
                  "key",
                  "value"
                ],
                "additionalProperties": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session",
        "description": "管理画面セッションCookie"
      },
      "patientCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "patient_id",
        "description": "患者IDCookie（LINE認証後に自動付与）"
      },
      "platformCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "platform_session",
        "description": "プラットフォーム管理セッションCookie"
      }
    }
  },
...
}